ppp-processor/requirements.txt
requests>=2.28.0
pyyaml>=6.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
Pillow>=9.5.0
numpy>=1.24.0

# PPP Processor v2.0 additions
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
celery[redis]>=5.3.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
websockets>=12.0
httpx>=0.25.0
scikit-image>=0.22.0
